+++ examples/dao/test/dao-app-flow.js
@@ -2,10 +2,10 @@ const { convert } = require('@algo-builder/algob');
 const { Runtime, AccountStore } = require('@algo-builder/runtime');
 const { types, parsing } = require('@algo-builder/web');
 const { assert } = require('chai');
-const { ProposalType, Vote } = require('../scripts/run/common/common');
+const { ProposalType, Vote, ExampleProposalConfig, DAOActions } = require('../scripts/run/common/common');
 const {
   now, mkProposalTx, mkDepositVoteTokenTx, mkWithdrawVoteDepositTx,
-  mkClearVoteRecordTx, mkClearProposalTx
+  mkClearVoteRecordTx, mkClearProposalTx, votingStart, votingEnd, executeBefore
 } = require('../scripts/run/common/tx-params');
 
 const minBalance = 10e6; // 10 ALGO's
@@ -18,7 +18,7 @@ const initialBalance = 200e6;
  *    2. Create DAO App
  *    3. Compile & fund lsig's (deposit, daoFund, proposal)
  *    4. Add depositLsig address to DAO app
- *    Note: Gov tokens are holded by proposer, voterA, voterB
+ *    Note: Gov tokens are holded by proposerA, voterA, voterB
  * + add proposal
  * + deposit vote tokens
  * + register vote (by token holders voterA & voterB)
@@ -31,12 +31,14 @@ const initialBalance = 200e6;
 describe('DAO test', function () {
   const master = new AccountStore(1000e6);
   let creator = new AccountStore(initialBalance);
-  let proposer = new AccountStore(initialBalance);
+  let proposerA = new AccountStore(initialBalance);
+  let proposerB = new AccountStore(initialBalance);
   let voterA = new AccountStore(initialBalance);
   let voterB = new AccountStore(initialBalance);
   let depositLsigAcc = new AccountStore(initialBalance); // runtime.account of depositLsig.address()
   let daoFundLsigAcc = new AccountStore(initialBalance);
-  let proposalLsigAcc = new AccountStore(initialBalance);
+  let proposalALsigAcc = new AccountStore(initialBalance);
+  let proposalBLsigAcc = new AccountStore(initialBalance);
 
   let runtime;
   let appCreationFlags; // deploy app params (sender, storage schema)
@@ -44,12 +46,13 @@ describe('DAO test', function () {
   let govTokenID;
   let depositLsig;
   let daoFundLsig;
-  let proposalLsig;
+  let proposalALsig;
+  let proposalBLsig;
 
   this.beforeAll(async function () {
     runtime = new Runtime([
-      master, creator, proposer, voterA, voterB,
-      depositLsigAcc, daoFundLsigAcc, proposalLsigAcc
+      master, creator, proposerA, proposerB, voterA, voterB,
+      depositLsigAcc, daoFundLsigAcc, proposalALsigAcc, proposalBLsigAcc
     ]);
 
     appCreationFlags = {
@@ -66,12 +69,14 @@ describe('DAO test', function () {
   // fetch latest account state
   function syncAccounts () {
     creator = runtime.getAccount(creator.address);
-    proposer = runtime.getAccount(proposer.address);
+    proposerA = runtime.getAccount(proposerA.address);
+    proposerB = runtime.getAccount(proposerB.address);
     voterA = runtime.getAccount(voterA.address);
     voterB = runtime.getAccount(voterB.address);
     depositLsigAcc = runtime.getAccount(depositLsigAcc.address);
     daoFundLsigAcc = runtime.getAccount(daoFundLsigAcc.address);
-    proposalLsigAcc = runtime.getAccount(proposalLsigAcc.address);
+    proposalALsigAcc = runtime.getAccount(proposalALsigAcc.address);
+    proposalBLsigAcc = runtime.getAccount(proposalBLsigAcc.address);
   }
 
   // DAO App initialization parameters
@@ -81,37 +86,8 @@ describe('DAO test', function () {
   const maxDuration = 5 * 60; // 5min (maximum voting time in number of seconds)
   const url = 'www.my-url.com';
 
-  it('DAO flow test', () => {
-    /**
-    * Flow:
-    *
-    * Create Gov Token (ASA used to represent voting power)
-    * Deploy DAO App
-    * Compile & fund lsigs:
-    *   a) depositLsig: holds "vote token deposits" & "Proposal deposits"
-    *   b) proposalLsig: used by proposer account for adding proposal
-    *   c) daoFundLsig: represents DAO treasury
-    * Save deposit lsig address in DAO app (only callable by creator)
-    * Intial distribution of few gov token(s) to accounts (creator, proposer, voters, lsigs)
-    * Add proposal record in proposalLsig(as sender) + make deposit to deposit Lsig
-    * Deposit vote tokens: voterA deposit 6 tokens, voterB deposits 8 tokens.
-    * Voting (note: each token == 1 vote):
-    *   a) move latest time after voting_now & <= voting_end
-    *   b) voterA votes "yes" (== 6 votes)
-    *   c) voterA votes "abstain" (== 8 votes).
-    * Execute:
-    *   a) move latest time after voting_end & <= execute_before
-    *   b) Call to DAO app + Algo transfer tx (as per proposal instruction)
-    * Withdrawing deposit:
-    *   a) Call to DAO app + asset transfer (deposit_lsig => voter)
-    *   b) voterA withdraws his 6 votes, voterB withdraws his 8 votes.
-    * Clear Voting Record (by voterA & voterB)
-    * Clear proposal:
-    *   a) optIn to Gov Token by proposalLsig
-    *   b) Call to DAO app by proposalLsig + asset transfer transaction from depositLsig -> proposalLsig
-    */
-
-    govTokenID = runtime.deployASA(
+  function setUpDAO () {
+    govTokenID = runtime.addAsset(
       'gov-token', { creator: { ...creator.account, name: 'dao-creator' } }).assetID;
 
     const daoAppArgs = [
@@ -146,12 +122,16 @@ describe('DAO test', function () {
     daoFundLsig = runtime.loadLogic('dao-fund-lsig.py', scInitParam);
     daoFundLsigAcc = runtime.getAccount(daoFundLsig.address());
 
-    proposalLsig = runtime.loadLogic('proposal-lsig.py',
-      { ARG_OWNER: proposer.address, ARG_DAO_APP_ID: appID });
-    proposalLsigAcc = runtime.getAccount(proposalLsig.address());
+    proposalALsig = runtime.loadLogic('proposal-lsig.py',
+      { ARG_OWNER: proposerA.address, ARG_DAO_APP_ID: appID });
+    proposalALsigAcc = runtime.getAccount(proposalALsig.address());
+
+    proposalBLsig = runtime.loadLogic('proposal-lsig.py',
+      { ARG_OWNER: proposerB.address, ARG_DAO_APP_ID: appID });
+    proposalBLsigAcc = runtime.getAccount(proposalBLsig.address());
 
     // fund lsig's
-    for (const lsig of [depositLsig, daoFundLsig, proposalLsig]) {
+    for (const lsig of [depositLsig, daoFundLsig, proposalALsig, proposalBLsig]) {
       runtime.fundLsig(master.account, lsig.address(), minBalance + 10000);
     }
     syncAccounts();
@@ -182,13 +162,13 @@ describe('DAO test', function () {
     assert.deepEqual(getGlobal('deposit_lsig'), convert.addressToPk(depositLsig.address()));
 
     // optIn to ASA(Gov Token) by accounts
-    for (const acc of [proposer, voterA, voterB, depositLsigAcc, daoFundLsigAcc]) {
+    for (const acc of [proposerA, proposerB, voterA, voterB, depositLsigAcc, daoFundLsigAcc]) {
       runtime.optIntoASA(govTokenID, acc.address, {});
     }
     syncAccounts();
 
     // verify optIn
-    for (const acc of [proposer, voterA, voterB, depositLsigAcc, daoFundLsigAcc]) {
+    for (const acc of [proposerA, proposerB, voterA, voterB, depositLsigAcc, daoFundLsigAcc]) {
       assert.isDefined(acc.getAssetHolding(govTokenID));
     }
 
@@ -197,35 +177,70 @@ describe('DAO test', function () {
       type: types.TransactionType.TransferAsset,
       sign: types.SignType.SecretKey,
       fromAccount: creator.account,
-      amount: 100,
+      amount: 1000,
       assetID: govTokenID,
       payFlags: { totalFee: 1000 }
     };
 
     runtime.executeTx([
-      { ...distributeGovTokenParams, toAccountAddr: proposer.address },
+      { ...distributeGovTokenParams, toAccountAddr: proposerA.address },
+      { ...distributeGovTokenParams, toAccountAddr: proposerB.address },
       { ...distributeGovTokenParams, toAccountAddr: voterA.address },
       { ...distributeGovTokenParams, toAccountAddr: voterB.address }
     ]);
     syncAccounts();
 
-    assert.equal(proposer.getAssetHolding(govTokenID).amount, 100);
-    assert.equal(voterA.getAssetHolding(govTokenID).amount, 100);
-    assert.equal(voterB.getAssetHolding(govTokenID).amount, 100);
+    assert.equal(proposerA.getAssetHolding(govTokenID).amount, 1000);
+    assert.equal(proposerB.getAssetHolding(govTokenID).amount, 1000);
+    assert.equal(voterA.getAssetHolding(govTokenID).amount, 1000);
+    assert.equal(voterB.getAssetHolding(govTokenID).amount, 1000);
+  }
+
+  it('DAO flow test', () => {
+    /**
+    * Flow:
+    *
+    * Create Gov Token (ASA used to represent voting power)
+    * Deploy DAO App
+    * Compile & fund lsigs:
+    *   a) depositLsig: holds "vote token deposits" & "Proposal deposits"
+    *   b) proposalALsig: used by proposerA account for adding proposal
+    *   c) daoFundLsig: represents DAO treasury
+    * Save deposit lsig address in DAO app (only callable by creator)
+    * Intial distribution of few gov token(s) to accounts (creator, proposerA, voters, lsigs)
+    * Add proposal record in proposalALsig(as sender) + make deposit to deposit Lsig
+    * Deposit vote tokens: voterA deposit 6 tokens, voterB deposits 8 tokens.
+    * Voting (note: each token == 1 vote):
+    *   a) move latest time after voting_now & <= voting_end
+    *   b) voterA votes "yes" (== 6 votes)
+    *   c) voterA votes "abstain" (== 8 votes).
+    * Execute:
+    *   a) move latest time after voting_end & <= execute_before
+    *   b) Call to DAO app + Algo transfer tx (as per proposal instruction)
+    * Withdrawing deposit:
+    *   a) Call to DAO app + asset transfer (deposit_lsig => voter)
+    *   b) voterA withdraws his 6 votes, voterB withdraws his 8 votes.
+    * Clear Voting Record (by voterA & voterB)
+    * Clear proposal:
+    *   a) optIn to Gov Token by proposalALsig
+    *   b) Call to DAO app by proposalALsig + asset transfer transaction from depositLsig -> proposalALsig
+    */
+
+    setUpDAO();
 
     /* --------------------  Add proposal  -------------------- */
 
-    // optIn to DAO by proposalLsig
-    runtime.optInToApp(proposalLsig.address(), appID, {}, {});
+    // optIn to DAO by proposalALsig
+    runtime.optInToApp(proposalALsig.address(), appID, {}, {});
 
     let beforeBal = depositLsigAcc.getAssetHolding(govTokenID).amount;
 
     const addProposalTx = mkProposalTx(
       appID,
       govTokenID,
-      proposer.account,
+      proposerA.account,
       depositLsig,
-      proposalLsig,
+      proposalALsig,
       daoFundLsig
     );
 
@@ -237,18 +252,18 @@ describe('DAO test', function () {
     syncAccounts();
 
     // assert proposal config is added
-    assert.deepEqual(proposalLsigAcc.getLocalState(appID, 'name'), parsing.stringToBytes('my-custom-proposal'));
-    assert.deepEqual(proposalLsigAcc.getLocalState(appID, 'url'), parsing.stringToBytes('www.myurl.com'));
-    assert.deepEqual(proposalLsigAcc.getLocalState(appID, 'url_hash'), parsing.stringToBytes('url-hash'));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'name'), parsing.stringToBytes(ExampleProposalConfig.name));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'url'), parsing.stringToBytes(ExampleProposalConfig.URL));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'url_hash'), parsing.stringToBytes(ExampleProposalConfig.URLHash));
     // empty hash_algo must save sha256
-    assert.deepEqual(proposalLsigAcc.getLocalState(appID, 'hash_algo'), parsing.stringToBytes('sha256'));
-    assert.deepEqual(proposalLsigAcc.getLocalState(appID, 'voting_start'), BigInt(now + (1 * 60)));
-    assert.deepEqual(proposalLsigAcc.getLocalState(appID, 'voting_end'), BigInt(now + (3 * 60)));
-    assert.deepEqual(proposalLsigAcc.getLocalState(appID, 'execute_before'), BigInt(now + (7 * 60)));
-    assert.deepEqual(proposalLsigAcc.getLocalState(appID, 'type'), BigInt(ProposalType.ALGO_TRANSFER));
-    assert.deepEqual(proposalLsigAcc.getLocalState(appID, 'from'), parsing.addressToPk(daoFundLsig.address()));
-    assert.deepEqual(proposalLsigAcc.getLocalState(appID, 'recipient'), parsing.addressToPk(proposer.address));
-    assert.deepEqual(proposalLsigAcc.getLocalState(appID, 'amount'), BigInt(2e6));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'hash_algo'), parsing.stringToBytes('sha256'));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'voting_start'), BigInt(votingStart));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'voting_end'), BigInt(votingEnd));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'execute_before'), BigInt(executeBefore));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'type'), BigInt(ProposalType.ALGO_TRANSFER));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'from'), parsing.addressToPk(daoFundLsig.address()));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'recipient'), parsing.addressToPk(proposerA.address));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'amount'), BigInt(2e6));
 
     // verify deposit recieved in depositLsig
     assert.deepEqual(depositLsigAcc.getAssetHolding(govTokenID).amount, beforeBal + 15n);
@@ -289,7 +304,7 @@ describe('DAO test', function () {
     assert.deepEqual(depositLsigAcc.getAssetHolding(govTokenID).amount, beforeBal + 6n + 8n);
 
     /* --------------------  Vote  -------------------- */
-    runtime.setRoundAndTimestamp(10, now + (2 * 60));
+    runtime.setRoundAndTimestamp(10, Math.round((votingStart + votingEnd) / 2));
 
     // call to DAO app by voter (to register deposited votes)
     const registerVoteParam = {
@@ -297,14 +312,14 @@ describe('DAO test', function () {
       sign: types.SignType.SecretKey,
       appID: appID,
       payFlags: { totalFee: 2000 },
-      accounts: [proposalLsig.address()]
+      accounts: [proposalALsig.address()]
     };
 
     // voting by user A
     const registerVoteA = {
       ...registerVoteParam,
       fromAccount: voterA.account,
-      appArgs: ['str:register_vote', `str:${Vote.YES}`]
+      appArgs: [DAOActions.registerVote, `str:${Vote.YES}`]
     };
     runtime.executeTx(registerVoteA);
 
@@ -312,62 +327,62 @@ describe('DAO test', function () {
     const registerVoteB = {
       ...registerVoteParam,
       fromAccount: voterB.account,
-      appArgs: ['str:register_vote', `str:${Vote.ABSTAIN}`]
+      appArgs: [DAOActions.registerVote, `str:${Vote.ABSTAIN}`]
     };
     runtime.executeTx(registerVoteB);
     syncAccounts();
 
-    // concatination of "p_" & proposalLsig.address
-    const key = new Uint8Array([...parsing.stringToBytes('p_'), ...parsing.addressToPk(proposalLsig.address())]);
+    // concatination of "p_" & proposalALsig.address
+    const key = new Uint8Array([...parsing.stringToBytes('p_'), ...parsing.addressToPk(proposalALsig.address())]);
 
     // verify sender account has p_proposal set
     assert.isDefined(voterA.getLocalState(appID, key));
-    assert.isDefined(voterA.getLocalState(appID, key));
+    assert.isDefined(voterB.getLocalState(appID, key));
 
     // verify voting count
     // + voterA registered 6 "yes" votes
     // + voterB registered 8 "abstain" votes
-    assert.equal(proposalLsigAcc.getLocalState(appID, 'yes'), 6n);
-    assert.equal(proposalLsigAcc.getLocalState(appID, 'abstain'), 8n);
-    assert.isUndefined(proposalLsigAcc.getLocalState(appID, 'no')); // we didn't vote for "no"
+    assert.equal(proposalALsigAcc.getLocalState(appID, 'yes'), 6n);
+    assert.equal(proposalALsigAcc.getLocalState(appID, 'abstain'), 8n);
+    assert.isUndefined(proposalALsigAcc.getLocalState(appID, 'no')); // we didn't vote for "no"
 
     /* --------------------  Execute  -------------------- */
 
     runtime.setRoundAndTimestamp(15, now + (5 * 60));
 
     // verify proposal not executed before
-    assert.equal(proposalLsigAcc.getLocalState(appID, 'executed'), 0n);
+    assert.equal(proposalALsigAcc.getLocalState(appID, 'executed'), 0n);
 
-    const beforeProposerBal = proposer.balance();
+    const beforeProposerBal = proposerA.balance();
     const executeParams = [
       {
         type: types.TransactionType.CallApp,
         sign: types.SignType.SecretKey,
-        fromAccount: proposer.account,
+        fromAccount: proposerA.account,
         appID: appID,
         payFlags: { totalFee: 2000 },
-        appArgs: ['str:execute'],
-        accounts: [proposalLsig.address()]
+        appArgs: [DAOActions.execute],
+        accounts: [proposalALsig.address()]
       },
       // tx1 as per proposal instructions (set in ./add_proposal.js)
       {
         type: types.TransactionType.TransferAlgo,
         sign: types.SignType.LogicSignature,
         fromAccountAddr: daoFundLsig.address(),
-        toAccountAddr: proposer.address,
+        toAccountAddr: proposerA.address,
         amountMicroAlgos: 2e6,
         lsig: daoFundLsig,
-        payFlags: { totalFee: 0 } // fee must be paid by proposer
+        payFlags: { totalFee: 0 } // fee must be paid by proposerA
       }
     ];
     runtime.executeTx(executeParams);
     syncAccounts();
 
     // verify executed is set to be true
-    assert.equal(proposalLsigAcc.getLocalState(appID, 'executed'), 1n);
+    assert.equal(proposalALsigAcc.getLocalState(appID, 'executed'), 1n);
 
     // verify payment recieved from dao fund
-    assert.equal(proposer.balance(), beforeProposerBal + BigInt(2e6) - 2000n);
+    assert.equal(proposerA.balance(), beforeProposerBal + BigInt(2e6) - 2000n);
 
     /* --------------------  Withdraw Vote Deposit  -------------------- */
 
@@ -417,7 +432,7 @@ describe('DAO test', function () {
     const clearRecordVoterA = mkClearVoteRecordTx(
       appID,
       voterA.account,
-      proposalLsig.address()
+      proposalALsig.address()
     );
     runtime.executeTx(clearRecordVoterA);
 
@@ -425,7 +440,7 @@ describe('DAO test', function () {
     const clearRecordVoterB = mkClearVoteRecordTx(
       appID,
       voterB.account,
-      proposalLsig.address()
+      proposalALsig.address()
     );
     runtime.executeTx(clearRecordVoterB);
     syncAccounts();
@@ -436,21 +451,21 @@ describe('DAO test', function () {
 
     /* --------------------  Clear Proposal  -------------------- */
 
-    // optIn to GovToken by proposalLsig (protected by proposer account using opt-in lock)
+    // optIn to GovToken by proposalALsig (protected by proposerA account using opt-in lock)
     const optInTx = [
       {
         type: types.TransactionType.TransferAlgo,
         sign: types.SignType.SecretKey,
-        fromAccount: proposer.account,
-        toAccountAddr: proposalLsig.address(),
+        fromAccount: proposerA.account,
+        toAccountAddr: proposalALsig.address(),
         amountMicroAlgos: 0,
         payFlags: {}
       },
       {
         type: types.TransactionType.OptInASA,
         sign: types.SignType.LogicSignature,
-        fromAccountAddr: proposalLsig.address(),
-        lsig: proposalLsig,
+        fromAccountAddr: proposalALsig.address(),
+        lsig: proposalALsig,
         assetID: govTokenID,
         payFlags: {}
       }
@@ -458,20 +473,20 @@ describe('DAO test', function () {
     runtime.executeTx(optInTx);
     syncAccounts();
 
-    assert.isDefined(proposalLsigAcc.getAssetHolding(govTokenID));
+    assert.isDefined(proposalALsigAcc.getAssetHolding(govTokenID));
 
     const clearProposalParam = mkClearProposalTx(
       appID,
       govTokenID,
       depositLsig,
-      proposalLsig,
+      proposalALsig,
       15 // set as deposit in DAO App
     );
     runtime.executeTx(clearProposalParam);
     syncAccounts();
 
-    // verify proposalLsig recieved back deposit of 15 tokens
-    assert.equal(proposalLsigAcc.getAssetHolding(govTokenID).amount, 15n);
+    // verify proposalALsig recieved back deposit of 15 tokens
+    assert.equal(proposalALsigAcc.getAssetHolding(govTokenID).amount, 15n);
 
     // verify proposal config is deleted from localstate
     for (
@@ -479,7 +494,377 @@ describe('DAO test', function () {
         'voting_end', 'execute_before', 'type', 'from', 'recipient',
         'amount', 'yes', 'no', 'abstain']
     ) {
-      assert.isUndefined(proposalLsigAcc.getLocalState(appID, key));
+      assert.isUndefined(proposalALsigAcc.getLocalState(appID, key));
+    }
+  });
+
+  it('Should not allow to vote again with newly locked tokens', () => {
+    /**
+    * Flow:
+    * + Setup DAO
+    * + Add proposalA
+    * + Lock 100 tokens by voterA (pass)
+    * + Vote by voterA (pass)
+    * + Lock 50 more tokens by voterA (pass)
+    * + VoterA tries to voter again with newly locked tokens (fail)
+    */
+
+    setUpDAO();
+
+    /* --------------------  Add proposal  -------------------- */
+
+    // optIn to DAO by proposalALsig
+    runtime.optInToApp(proposalALsig.address(), appID, {}, {});
+
+    let beforeBal = depositLsigAcc.getAssetHolding(govTokenID).amount;
+
+    const addProposalTx = mkProposalTx(
+      appID,
+      govTokenID,
+      proposerA.account,
+      depositLsig,
+      proposalALsig,
+      daoFundLsig
+    );
+
+    // set time (after now)
+    runtime.setRoundAndTimestamp(5, now + 10);
+
+    addProposalTx[1].amount = 15;
+    runtime.executeTx(addProposalTx);
+    syncAccounts();
+
+    // assert proposal config is added
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'name'), parsing.stringToBytes(ExampleProposalConfig.name));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'url'), parsing.stringToBytes(ExampleProposalConfig.URL));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'url_hash'), parsing.stringToBytes(ExampleProposalConfig.URLHash));
+    // empty hash_algo must save sha256
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'hash_algo'), parsing.stringToBytes('sha256'));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'voting_start'), BigInt(votingStart));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'voting_end'), BigInt(votingEnd));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'execute_before'), BigInt(executeBefore));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'type'), BigInt(ProposalType.ALGO_TRANSFER));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'from'), parsing.addressToPk(daoFundLsig.address()));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'recipient'), parsing.addressToPk(proposerA.address));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'amount'), BigInt(2e6));
+
+    // verify deposit recieved in depositLsig
+    assert.deepEqual(depositLsigAcc.getAssetHolding(govTokenID).amount, beforeBal + 15n);
+
+    /* --------------------  Deposit Vote Token (Lock 100 tokens by A)  -------------------- */
+
+    // optIn to DAO by voterA & voterB
+    runtime.optInToApp(voterA.address, appID, {}, {});
+    beforeBal = depositLsigAcc.getAssetHolding(govTokenID).amount;
+
+    // lock 100 votes by voterA
+    const depositVoteParamA = mkDepositVoteTokenTx(
+      appID,
+      govTokenID,
+      voterA.account,
+      depositLsig,
+      100
+    );
+    runtime.executeTx(depositVoteParamA);
+    syncAccounts();
+
+    // verify sender.deposit is set
+    assert.deepEqual(voterA.getLocalState(appID, 'deposit'), 100n);
+    // verify 100 votes deposited
+    assert.deepEqual(depositLsigAcc.getAssetHolding(govTokenID).amount, beforeBal + 100n);
+
+    /* --------------------  Register 100 Votes by A  -------------------- */
+    runtime.setRoundAndTimestamp(10, Math.round((votingStart + votingEnd) / 2));
+
+    // call to DAO app by voter (to register deposited votes)
+    const registerVoteParam = {
+      type: types.TransactionType.CallApp,
+      sign: types.SignType.SecretKey,
+      fromAccount: voterA.account,
+      appID: appID,
+      payFlags: { totalFee: 2000 },
+      appArgs: [DAOActions.registerVote, `str:${Vote.YES}`],
+      accounts: [proposalALsig.address()]
+    };
+    runtime.executeTx(registerVoteParam);
+    syncAccounts();
+
+    // concatination of "p_" & proposalALsig.address
+    const key = new Uint8Array([...parsing.stringToBytes('p_'), ...parsing.addressToPk(proposalALsig.address())]);
+
+    // verify sender account has p_proposal set
+    assert.isDefined(voterA.getLocalState(appID, key));
+
+    // verify voting count
+    assert.equal(proposalALsigAcc.getLocalState(appID, 'yes'), 100n);
+    assert.isUndefined(proposalALsigAcc.getLocalState(appID, 'abstain')); // we didn't vote for "abstain"
+    assert.isUndefined(proposalALsigAcc.getLocalState(appID, 'no')); // we didn't vote for "no"
+
+    /* --------------------  Deposit Vote Token (Lock 50 tokens again by A)  -------------------- */
+    beforeBal = depositLsigAcc.getAssetHolding(govTokenID).amount;
+
+    // lock 50 votes by voterA
+    const depositVoteParam = mkDepositVoteTokenTx(
+      appID,
+      govTokenID,
+      voterA.account,
+      depositLsig,
+      50
+    );
+    runtime.executeTx(depositVoteParam);
+    syncAccounts();
+
+    // verify sender.deposit is set
+    assert.deepEqual(voterA.getLocalState(appID, 'deposit'), 150n); // 100 + 50
+    // verify 50 votes deposited
+    assert.deepEqual(depositLsigAcc.getAssetHolding(govTokenID).amount, beforeBal + 50n);
+
+    /* --------------------  Register 50 Votes by A (fails this time)  -------------------- */
+    assert.throws(
+      () => runtime.executeTx(registerVoteParam),
+      'RUNTIME_ERR1009: TEAL runtime encountered err opcode'
+    );
+  });
+
+  it('Should allow to vote again with newly locked tokens for different proposal', () => {
+    /**
+    * Flow:
+    * + Setup DAO
+    * + Add proposalA
+    * + Add proposalB
+    * + Lock 50 tokens by voterA (pass)
+    * + Vote by voterA (pass)
+    * + Lock 50 more tokens by voterA (pass)
+    * + VoterA tries to vote for proposalB (passes)
+    */
+
+    setUpDAO();
+
+    /* --------------------  Add proposal(s)  -------------------- */
+
+    // optIn to DAO by proposalALsig & proposalBLsig
+    runtime.optInToApp(proposalALsig.address(), appID, {}, {});
+    runtime.optInToApp(proposalBLsig.address(), appID, {}, {});
+
+    let beforeBal = depositLsigAcc.getAssetHolding(govTokenID).amount;
+
+    const addProposalATx = mkProposalTx(
+      appID,
+      govTokenID,
+      proposerA.account,
+      depositLsig,
+      proposalALsig,
+      daoFundLsig
+    );
+
+    const addProposalBTx = mkProposalTx(
+      appID,
+      govTokenID,
+      proposerB.account,
+      depositLsig,
+      proposalBLsig,
+      daoFundLsig
+    );
+
+    // set time (after now)
+    runtime.setRoundAndTimestamp(5, now + 10);
+
+    addProposalATx[1].amount = 15;
+    addProposalBTx[1].amount = 15;
+    runtime.executeTx(addProposalATx);
+    runtime.executeTx(addProposalBTx);
+
+    syncAccounts();
+
+    // verify deposit recieved in depositLsig
+    assert.deepEqual(depositLsigAcc.getAssetHolding(govTokenID).amount, beforeBal + 15n + 15n);
+
+    /* --------------------  Deposit Vote Token (Lock 50 tokens by A)  -------------------- */
+
+    // optIn to DAO by voterA & voterB
+    runtime.optInToApp(voterA.address, appID, {}, {});
+    beforeBal = depositLsigAcc.getAssetHolding(govTokenID).amount;
+
+    // lock 50 votes by voterA
+    const depositVoteParamA = mkDepositVoteTokenTx(
+      appID,
+      govTokenID,
+      voterA.account,
+      depositLsig,
+      50
+    );
+    runtime.executeTx(depositVoteParamA);
+    syncAccounts();
+
+    // verify sender.deposit is set
+    assert.deepEqual(voterA.getLocalState(appID, 'deposit'), 50n);
+    // verify 50 votes deposited
+    assert.deepEqual(depositLsigAcc.getAssetHolding(govTokenID).amount, beforeBal + 50n);
+
+    /* --------------------  Register 50 Votes by A  -------------------- */
+    runtime.setRoundAndTimestamp(10, Math.round((votingStart + votingEnd) / 2));
+
+    // call to DAO app by voter (to register deposited votes)
+    const registerVoteParam = {
+      type: types.TransactionType.CallApp,
+      sign: types.SignType.SecretKey,
+      fromAccount: voterA.account,
+      appID: appID,
+      payFlags: { totalFee: 2000 },
+      appArgs: [DAOActions.registerVote, `str:${Vote.ABSTAIN}`],
+      accounts: [proposalALsig.address()]
+    };
+    runtime.executeTx(registerVoteParam);
+    syncAccounts();
+
+    // concatination of "p_" & proposalALsig.address
+    const key = new Uint8Array([...parsing.stringToBytes('p_'), ...parsing.addressToPk(proposalALsig.address())]);
+
+    // verify sender account has p_proposal set
+    assert.isDefined(voterA.getLocalState(appID, key));
+
+    // verify voting count
+    assert.equal(proposalALsigAcc.getLocalState(appID, 'abstain'), 50n);
+    assert.isUndefined(proposalALsigAcc.getLocalState(appID, 'yes')); // we didn't vote for "yes"
+    assert.isUndefined(proposalALsigAcc.getLocalState(appID, 'no')); // we didn't vote for "no"
+
+    /* --------------------  Deposit Vote Token (Lock 50 tokens again by A)  -------------------- */
+    beforeBal = depositLsigAcc.getAssetHolding(govTokenID).amount;
+
+    // lock 50 votes by voterA
+    const depositVoteParam = mkDepositVoteTokenTx(
+      appID,
+      govTokenID,
+      voterA.account,
+      depositLsig,
+      50
+    );
+    runtime.executeTx(depositVoteParam);
+    syncAccounts();
+
+    // verify sender.deposit is set
+    assert.deepEqual(voterA.getLocalState(appID, 'deposit'), 100n); // 50 + 50
+    // verify 50 votes deposited
+    assert.deepEqual(depositLsigAcc.getAssetHolding(govTokenID).amount, beforeBal + 50n);
+
+    /* --------------------  Register 100 Votes by A for proposalB (passes this time)  -------------------- */
+    // call to DAO app by voter (to register deposited votes)
+    const registerVoteParamForProposalB = {
+      ...registerVoteParam,
+      accounts: [proposalBLsig.address()]
+    };
+    runtime.executeTx(registerVoteParamForProposalB);
+    syncAccounts();
+
+    // concatination of "p_" & proposalBLsig.address
+    const proposalBKey = new Uint8Array([...parsing.stringToBytes('p_'), ...parsing.addressToPk(proposalBLsig.address())]);
+
+    // verify sender account has p_proposal set
+    assert.isDefined(voterA.getLocalState(appID, proposalBKey));
+
+    // verify voting count
+    assert.equal(proposalBLsigAcc.getLocalState(appID, 'abstain'), 100n);
+    assert.isUndefined(proposalBLsigAcc.getLocalState(appID, 'yes')); // we didn't vote for "yes"
+    assert.isUndefined(proposalBLsigAcc.getLocalState(appID, 'no')); // we didn't vote for "no"
+  });
+
+  it('Should allow to clear proposal if it is passsed execution', () => {
+    /**
+    * Flow:
+    * + Setup DAO
+    * + Add proposalA
+    * + Clear proposalA (passes if past execution)
+    */
+    setUpDAO();
+
+    /* --------------------  Add proposal  -------------------- */
+
+    // optIn to DAO by proposalALsig
+    runtime.optInToApp(proposalALsig.address(), appID, {}, {});
+
+    const beforeBal = depositLsigAcc.getAssetHolding(govTokenID).amount;
+
+    const addProposalTx = mkProposalTx(
+      appID,
+      govTokenID,
+      proposerA.account,
+      depositLsig,
+      proposalALsig,
+      daoFundLsig
+    );
+
+    // set time (after now)
+    runtime.setRoundAndTimestamp(5, now + 10);
+
+    addProposalTx[1].amount = 15;
+    runtime.executeTx(addProposalTx);
+    syncAccounts();
+
+    // assert proposal config is added
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'name'), parsing.stringToBytes(ExampleProposalConfig.name));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'url'), parsing.stringToBytes(ExampleProposalConfig.URL));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'url_hash'), parsing.stringToBytes(ExampleProposalConfig.URLHash));
+    // empty hash_algo must save sha256
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'hash_algo'), parsing.stringToBytes('sha256'));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'voting_start'), BigInt(votingStart));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'voting_end'), BigInt(votingEnd));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'execute_before'), BigInt(executeBefore));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'type'), BigInt(ProposalType.ALGO_TRANSFER));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'from'), parsing.addressToPk(daoFundLsig.address()));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'recipient'), parsing.addressToPk(proposerA.address));
+    assert.deepEqual(proposalALsigAcc.getLocalState(appID, 'amount'), BigInt(2e6));
+
+    // verify deposit recieved in depositLsig
+    assert.deepEqual(depositLsigAcc.getAssetHolding(govTokenID).amount, beforeBal + 15n);
+
+    /* --------------------  Clear Proposal  -------------------- */
+    // set time past executeBefore
+    runtime.setRoundAndTimestamp(5, executeBefore + 10);
+
+    // optIn to GovToken by proposalALsig (protected by proposerA account using opt-in lock)
+    const optInTx = [
+      {
+        type: types.TransactionType.TransferAlgo,
+        sign: types.SignType.SecretKey,
+        fromAccount: proposerA.account,
+        toAccountAddr: proposalALsig.address(),
+        amountMicroAlgos: 0,
+        payFlags: {}
+      },
+      {
+        type: types.TransactionType.OptInASA,
+        sign: types.SignType.LogicSignature,
+        fromAccountAddr: proposalALsig.address(),
+        lsig: proposalALsig,
+        assetID: govTokenID,
+        payFlags: {}
+      }
+    ];
+    runtime.executeTx(optInTx);
+    syncAccounts();
+
+    assert.isDefined(proposalALsigAcc.getAssetHolding(govTokenID));
+
+    const clearProposalParam = mkClearProposalTx(
+      appID,
+      govTokenID,
+      depositLsig,
+      proposalALsig,
+      15 // set as deposit in DAO App
+    );
+    runtime.executeTx(clearProposalParam);
+    syncAccounts();
+
+    // verify proposalALsig recieved back deposit of 15 tokens
+    assert.equal(proposalALsigAcc.getAssetHolding(govTokenID).amount, 15n);
+
+    // verify proposal config is deleted from localstate
+    for (
+      const key of ['name', 'url', 'url_hash', 'hash_algo', 'voting_start',
+        'voting_end', 'execute_before', 'type', 'from', 'recipient',
+        'amount', 'yes', 'no', 'abstain']
+    ) {
+      assert.isUndefined(proposalALsigAcc.getLocalState(appID, key));
     }
   });
 });